Refatoração dos testes, e inclusão de teste para função computação AFD
