+++ README.md
@@ -2,17 +2,48 @@
 
 Conjunto de fun&ccedil;&otilde;es em Python para simular os conceitos da disciplina de Teoria da Computa&ccedil;&atilde;o. Novas fun&ccedil;&otilde;es e exemplos ser&atilde;o adicionados ao longo do semestre.
 
+
+
+## Organização dos diretórios
+
+- `./exemplos` : exemplos apresentados em aula construído com as funções;
+
+- `./tc` : funções com os conceitos;
+
+- `./testes`: testes unitários para as funções construidas.
+
+
+
 ## Como executar
 
-### No Terminal
+#### No Terminal
 
 Para executar o programa no terminal, copie o exemplo para a pasta ra&iacute;z do projeto, e execute usando o seguinte comando:
-<p><code>python3 XXXX.py</code></p>
+
+```bash
+python3 XXXX.py
+```
+
 Sendo XXXX o nome do arquivo em que est&atilde;o os dados do exemplo a ser executado.
+
+
 
 ## Como testar
 
-### No Terminal
+### Requisitos
+- `pystest`
 
-Executar o arquivo unit.py na pasta ra&iacute;z do projeto, usando o seguinte comando:
-<p><code>pytest unit.py</code></p> 
+#### No Terminal
+
+Executar o comando abaixo na pasta ra&iacute;z do projeto:
+
+```bash
+python -m pytest testes -vv
+```
+
+Ou execute o script `unit_tests.py` com o comando abaixo:
+
+```bash
+python unit_tests.py
+```
+
