+++ testes/AFD/test_afd_aceita.py
@@ -1,0 +1,5 @@
+import tc.afd as tc
+from exemplos.afd_m1 import M1
+
+def test_aceita():
+    assert tc.aceita(M1, [1,1,0,1])
