+++ unit_tests.py
@@ -1,0 +1,3 @@
+import os
+
+os.system('python -m pytest testes -vv')
